refactor(core): extract controller and view to their own files
