+++ addon/components/debug-tools.ts
@@ -7,9 +7,9 @@ import { xml } from '@codemirror/lang-xml';
 import { html } from '@codemirror/lang-html';
 import sampleData from '../config/sample-data';
 import { EditorState } from '@codemirror/state';
-import { SayController } from '@lblod/ember-rdfa-editor/core/say-editor';
 import { unwrap } from '@lblod/ember-rdfa-editor/utils/option';
 import ApplicationInstance from '@ember/application/instance';
+import SayController from '@lblod/ember-rdfa-editor/core/say-controller';
 
 interface DebugToolArgs {
   controller?: SayController;
